// frontend/src/client/types.gen.ts
export type ChallengeCreate = {
    title: string;
    author: string;
    url: string;
    description: string;
    date_posted: string;
    is_active?: boolean;
};